+++ Managed Objects/TSSTPage.m
@@ -72,7 +72,7 @@ static NSSize monospaceCharacterSize;
     NSTextTab *tabStop;
     NSMutableArray *tabStops = [NSMutableArray array];
     int tabSize;
-    float tabLocation;
+    CGFloat tabLocation;
     /* Loop through the tab stops */
     for (tabSize = 8; tabSize < 120; tabSize += 8) {
         tabLocation = tabSize * monospaceCharacterSize.width;
@@ -117,8 +117,8 @@ static NSSize monospaceCharacterSize;
         return YES;
     }
 
-    float defaultAspect = 1;
-    float aspect = [[self valueForKey:@"aspectRatio"] floatValue];
+    CGFloat defaultAspect = 1;
+    CGFloat aspect = [[self valueForKey:@"aspectRatio"] floatValue];
     if (!aspect) {
         NSData *imageData = [self pageData];
         [self setOwnSizeInfoWithData:imageData];
@@ -130,7 +130,7 @@ static NSSize monospaceCharacterSize;
 
 - (void)setOwnSizeInfoWithData:(NSData *)imageData
 {
-    float aspect;
+    CGFloat aspect;
     NSSize imageSize;
     NSBitmapImageRep *pageRep = [NSBitmapImageRep imageRepWithData:imageData];
     imageSize = NSMakeSize([pageRep pixelsWide], [pageRep pixelsHigh]);
